+++ maibot_sdk/__init__.py
@@ -16,7 +16,7 @@ from .llm_provider import LLMProviderBase
 from .plugin import MaiBotPlugin
 from .types import CONFIG_RELOAD_SCOPE_SELF, ON_BOT_CONFIG_RELOAD, ON_MODEL_CONFIG_RELOAD
 
-__version__ = "2.4.0"
+__version__ = "2.5.0"
 
 __all__ = [
     "MaiBotPlugin",